+++ backend/canisters/multi_user/impl/src/queries/contacts.rs
@@ -5,19 +5,5 @@ use user_canister::contacts::{Response::*, *};
 
 #[query(guard = "caller_is_hosted_user", msgpack = true)]
 fn contacts(_args: Args) -> Response {
-    read_state(|state| {
-        state.with_caller_user(|_, user| {
-            Success(SuccessResult {
-                contacts: user
-                    .contacts
-                    .all()
-                    .into_iter()
-                    .map(|(user_id, contact)| Contact {
-                        user_id,
-                        nickname: contact.nickname,
-                    })
-                    .collect(),
-            })
-        })
-    })
+    read_state(|state| state.with_caller_user(|_, user| Success(user_core::queries::contacts(user))))
 }