+++ backend/canisters/user/impl/src/queries/contacts.rs
@@ -5,19 +5,5 @@ use user_canister::contacts::{Response::*, *};
 
 #[query(guard = "caller_is_owner", msgpack = true)]
 fn contacts(_args: Args) -> Response {
-    read_state(|state| {
-        Success(SuccessResult {
-            contacts: state
-                .data
-                .user
-                .contacts
-                .all()
-                .into_iter()
-                .map(|(user_id, contact)| Contact {
-                    user_id,
-                    nickname: contact.nickname,
-                })
-                .collect(),
-        })
-    })
+    read_state(|state| Success(user_core::queries::contacts(&state.data.user)))
 }