+++ backend/canisters/user/impl/src/queries/messages_by_message_index.rs
@@ -2,7 +2,6 @@ use crate::guards::caller_is_owner;
 use crate::queries::check_replica_up_to_date;
 use crate::{RuntimeState, read_state};
 use canister_api_macros::query;
-use chat_events::Reader;
 use oc_error_codes::OCErrorCode;
 use types::{MessagesResponse, OCResult};
 use user_canister::messages_by_message_index::{Response::*, *};
@@ -20,23 +19,5 @@ fn messages_by_message_index_impl(args: Args, state: &RuntimeState) -> OCResult<
         return Err(OCErrorCode::ReplicaNotUpToDate.with_message(now));
     }
 
-    let chat = state.data.user.direct_chats.get_or_err(&args.user_id.into())?;
-    let Some(events_reader) = chat.events_reader(args.thread_root_message_index) else {
-        return Err(OCErrorCode::ThreadNotFound.into());
-    };
-
-    let my_user_id = state.env.canister_id().into();
-    let messages: Vec<_> = args
-        .messages
-        .into_iter()
-        .filter_map(|m| events_reader.message_event(m.into(), Some(my_user_id)))
-        .collect();
-    let latest_event_index = events_reader.latest_event_index().unwrap();
-    let chat_last_updated = chat.last_updated();
-
-    Ok(MessagesResponse {
-        messages,
-        latest_event_index,
-        chat_last_updated,
-    })
+    user_core::queries::messages_by_message_index(&state.data.user, args, state.env.canister_id().into())
 }